+++ django_cotton/tests/test_basic.py
@@ -102,124 +102,3 @@ class BasicComponentTests(CottonTestCase):
                 response,
                 """My template path was not specified in settings!""",
             )
-
-    def test_only_gives_isolated_context(self):
-        self.create_template(
-            "cotton/only.html",
-            """
-                <a class="{{ class|default:"donttouch" }}">test</a>
-            """,
-        )
-        self.create_template(
-            "no_only_view.html",
-            """
-            <c-only />
-            """,
-            "no_only_view/",
-            context={"class": "herebedragons"},  # this should pass to `class` successfully
-        )
-
-        with self.settings(ROOT_URLCONF=self.url_conf()):
-            response = self.client.get("/no_only_view/")
-            self.assertNotContains(response, "donttouch")
-            self.assertContains(response, "herebedragons")
-
-        self.create_template(
-            "only_view.html",
-            """
-            <c-only only />
-            """,
-            "only_view/",
-            context={
-                "class": "herebedragons"
-            },  # this should not pass to `class` due to `only` being present
-        )
-
-        with self.settings(ROOT_URLCONF=self.url_conf()):
-            response = self.client.get("/only_view/")
-            self.assertNotContains(response, "herebedragons")
-            self.assertContains(response, "donttouch")
-
-        self.create_template(
-            "only_view2.html",
-            """
-            <c-only class="october" only />
-            """,
-            "only_view2/",
-            context={
-                "class": "herebedragons"
-            },  # this should not pass to `class` due to `only` being present
-        )
-
-        with self.settings(ROOT_URLCONF=self.url_conf()):
-            response = self.client.get("/only_view2/")
-            self.assertNotContains(response, "herebedragons")
-            self.assertNotContains(response, "donttouch")
-            self.assertContains(response, "october")
-
-    def test_only_with_dynamic_components(self):
-        self.create_template(
-            "cotton/dynamic_only.html",
-            """
-            From parent comp scope: '{{ class }}'
-            From view context scope: '{{ view_item }}'
-            Direct attribute: '{{ direct }}'
-            """,
-        )
-
-        self.create_template(
-            "cotton/middle_component.html",
-            """
-            <c-component is="{{ comp }}" only direct="yes" />
-            """,
-        )
-
-        self.create_template(
-            "dynamic_only_view.html",
-            """<c-middle-component class="mb-5" />""",
-            "view/",
-            context={"comp": "dynamic_only", "view_item": "blee"},
-        )
-
-        with self.settings(ROOT_URLCONF=self.url_conf()):
-            response = self.client.get("/view/")
-            self.assertContains(response, "From parent comp scope: ''")
-            self.assertContains(response, "From view context scope: ''")
-            self.assertContains(response, "Direct attribute: 'yes'")
-
-    def test_context_isolated_by_default(self):
-        self.create_template(
-            "cotton/receiver.html",
-            """
-            {{ global }} 
-            {{ direct }} 
-            {{ from_context_processor }}
-            
-            Some context from django builtins:
-            csrf: "{{ csrf_token }}" 
-            request: "{{ request }}"
-            messages: "{{ messages }}"
-            user: "{{ user }}"
-            perms: "{{ perms }}"
-            debug: "{{ debug }}"
-            """,
-        )
-
-        self.create_template(
-            "context_isolation_view.html",
-            """<c-receiver direct="hello" />""",
-            "view/",
-            context={"global": "shouldnotbeseen"},
-        )
-
-        # with example_processor added and 'logo' in the context
-
-        with self.settings(ROOT_URLCONF=self.url_conf()):
-            response = self.client.get("/view/")
-            print(response.content.decode())
-            self.assertNotContains(response, "shouldnotbeseen")
-            self.assertContains(response, "hello")
-            self.assertContains(response, "logo.png")
-            self.assertNotContains(response, 'csrf: ""')
-            self.assertContains(response, 'user: "AnonymousUser"')
-            self.assertNotContains(response, 'perms: ""')
