+++ docs/docs_project/docs_project/settings.py
@@ -70,7 +70,6 @@ TEMPLATES = [
                 "django.template.context_processors.debug",
                 "django.template.context_processors.request",
                 "django.contrib.auth.context_processors.auth",
-                "django.contrib.messages.context_processors.messages",
             ],
             "builtins": [
                 "django.templatetags.static",
