+++ docs/docs_project/docs_project/templates/configuration.html
@@ -6,6 +6,12 @@
     <h4>COTTON_DIR</h4>
     <p>str (default: 'cotton')</p>
 
-    <p>Change the default path in your templates directory where cotton components can be placed, for example "components".</p>
+    <c-hr />
+
+    <h4>COTTON_CONTEXT_ISOLATION_ENABLED</h4>
+    <p>str (default: 'True')</p>
+
+    <p>Set to `False` to allow global context to be available through all components. (see <a href="{% url 'components' %}#context-isolation">context isolation</a> for more information.</p>
+
 
 </c-layouts.with-sidebar>
